// ClientApp/src/app/views/employee/model/employees.module.ts
export class EmployeesModule {
  empId: number = 0;
  title: string = '';
  empEngName: string = '';
  empBdName: string = '';
  empCode: string = '';
  empGovNo: string = '';
  personalFileNo: string = '';
  dateOfBirth: Date | undefined;
  nid: number = 0;
  smartNid: number = 0;
  probabilityRetirementDate: Date | undefined;
  prlDate: Date | undefined;
  aspNetUserId: number = 0;
  employeeTypeId: number = 0;
  shiftId: number = 0;
  genderId: number = 0;
  maritalStatusId: number = 0;
  bloodGroupId: number = 0;
  nationalatyId: number = 0;
  mabileNumber: number = 0;
  email: string = '';
  birthRegNo: number = 0;
  placeOfBirth: string = '';
  freedomfighterStatus: string = '';
  religionId: number = 0;
  tinNo: number = 0;
  hairColorId: number = 0;
  eyesColor: number = 0;
  height: string = '';
  weight: string = '';
  healthIssueStatusId: number = 0;
  drivingLicenceNo: number = 0;
  drivingLicenceDate: Date | undefined;
  passportNo: string = '';
  passportExpireDate: Date | undefined;
  poolId: number = 0;
  officeId: number = 0;
  departmentId: number = 0;
  branchId: number = 0;
  subBranchId: number = 0;
  sectionId: number = 0;
  presentDesignationId: number = 0;
  presentGradeId: number = 0;
  basicPay: number = 0;
  joiningDate: Date | undefined;
  placeOfFirstJoining: string = '';
  firstGradeId: number = 0;
  firstScaleId: number = 0;
  retirementYear: number = 0;
  serviceLength: number = 0;
  serviceStatus: boolean = false;
  presentHouse: string = '';
  presentRoadOrVill: string = '';
  presentCountryId: number = 0;
  presentDivisionId: number = 0;
  presentDistrictId: number = 0;
  presentUpazilaId: number = 0;
  presentThanaId: number = 0;
  presentUnionId: number = 0;
  presentWardId: number = 0;
  permanentHouse: string = '';
  permanentRoadOrVill: string = '';
  permanentCountryId: number = 0;
  permanentDivisionId: number = 0;
  permanentDistrictId: number = 0;
  permanentUpazilaId: number = 0;
  permanentThanaId: number = 0;
  permanentUnionId: number = 0;
  permanentWardId: number = 0;
  permanentContractNo: number = 0;
  menuPosition: number = 0;
  isActive: boolean = true;
}